Organize module structure.
