+++ static/js/forms/index.js
@@ -1,0 +1,7 @@
+import a11y from "./a11y";
+import conditional from "./conditional";
+
+export default {
+  a11y,
+  conditional,
+};
