+++ templates/common/map.html
@@ -10,7 +10,7 @@
 </div>
 <script>
   window.addEventListener("DOMContentLoaded", function () {
-    a4a.initAppMap(
+    a4a.geo.initAppMap(
       {% if commune_json %}{{ commune_json | safe }}{% else %}null{% endif %},
       {% if erp %}{{ erp.pk }}{% else %}null{% endif %},
       {{ around | default_if_none:"null" }},
